+++ R/collinear.R
@@ -12,6 +12,8 @@
 #'
 #' Accepts a parallelization setup via [future::plan()] and a progress bar via [progressr::handlers()] (see examples).
 #'
+#' Unlike all other package functions, `collinear()` accepts a character vector as input for the argument `response`. When more than one response is provided, the output is a named list, with each element named after the given response contains a character vector of filtered predictors.
+#'
 #' @section Target Encoding:
 #'
 #' When the argument `response` names a **numeric response variable**, categorical predictors in `predictors` (or in the columns of `df` if `predictors` is NULL) are converted to numeric via **target encoding** with the function [target_encoding_lab()]. When `response` is NULL or names a categorical variable, target-encoding is skipped. This feature facilitates multicollinearity filtering in data frames with mixed column types.
@@ -37,27 +39,27 @@
 #'
 #' @section VIF-based Filtering:
 #'
-#' The function [vif_select()] computes Variance Inflation Factors and removes variables iteratively, until all variables in the resulting selection have a VIF below `max_vif`.
+#' The function [vif_select()] computes Variance Inflation Factors and removes variables iteratively, until all variables in the resulting selection have a VIF below `vif_max`.
 #'
-#' If the argument `preference_order` is not provided, all variables are ranked from lower to higher VIF, as returned by [vif_df()], and the variable with the higher VIF above `max_vif` is removed on each iteration.
+#' If the argument `preference_order` is not provided, all variables are ranked from lower to higher VIF, as returned by [vif_df()], and the variable with the higher VIF above `vif_max` is removed on each iteration.
 #'
-#' If `preference_order` is defined, whenever two or more variables are above `max_vif`, the one higher in `preference_order` is preserved, and the next one with a higher VIF is removed. For example, for the predictors and preference order \eqn{a} and \eqn{b}, if any of their VIFs is higher than `max_vif`, then \eqn{b} will be removed no matter whether its VIF is lower or higher than \eqn{a}'s VIF. If their VIF scores are lower than `max_vif`, then both are preserved.
+#' If `preference_order` is defined, whenever two or more variables are above `vif_max`, the one higher in `preference_order` is preserved, and the next one with a higher VIF is removed. For example, for the predictors and preference order \eqn{a} and \eqn{b}, if any of their VIFs is higher than `vif_max`, then \eqn{b} will be removed no matter whether its VIF is lower or higher than \eqn{a}'s VIF. If their VIF scores are lower than `vif_max`, then both are preserved.
 #'
 #' @section Pairwise Correlation Filtering:
 #'
-#' The function [cor_select()] applies a recursive forward selection algorithm to keep predictors with a maximum Pearson correlation with all other selected predictors lower than `max_cor`.
+#' The function [cor_select()] applies a recursive forward selection algorithm to keep predictors with a maximum Pearson correlation with all other selected predictors lower than `cor_max`.
 #'
 #' If the argument `preference_order` is not provided, the predictors are ranked from lower to higher sum of absolute pairwise correlation with all other predictors.
 #'
-#' If `preference_order` is defined, whenever two or more variables are above `max_cor`, the one higher in `preference_order` is preserved. For example, for the predictors and preference order \eqn{a} and \eqn{b}, if their correlation is higher than `max_cor`, then \eqn{b} will be removed and \eqn{a} preserved. If their correlation is lower than `max_cor`, then both are preserved.
+#' If `preference_order` is defined, whenever two or more variables are above `cor_max`, the one higher in `preference_order` is preserved. For example, for the predictors and preference order \eqn{a} and \eqn{b}, if their correlation is higher than `cor_max`, then \eqn{b} will be removed and \eqn{a} preserved. If their correlation is lower than `cor_max`, then both are preserved.
 #'
 #'
 #' @param df (required; data frame, tibble, or sf) A data frame with numeric predictors, and optionally a numeric response and categorical predictors. Default: NULL.
-#' @param response (optional, character string) Name of a numeric (non-numerics are ignored) response variable in `df`, only required if there are categorical variables named in the argument `predictors`. Default: NULL.
+#' @param response (optional, character string or vector) Name/s of response variable/s in `df`. Used in target encoding when it names a numeric variable, and in the computation of preference order if `preference_order = NULL`. Default: NULL.
 #' @param predictors (optional; character vector) Names of the variables to select from `df`. If omitted, all numeric columns in `df` are used instead. If argument `response` is not provided, non-numeric variables are ignored. Default: NULL
-#' @param encoding_method (optional; character string). Name of the target encoding method. One of: "mean" (default), "rank", "loo". If NULL, target encoding is disabled. Default: "mean"
+#' @param encoding_method (optional; character string). Name of the target encoding method. One of: "loo", "mean", or "rank". If NULL, target encoding is disabled. Default: "loo"
 #' @param preference_order (optional; character vector) Variable names in `predictors`. Defines a priority order, from first to last, to preserve variables during the selection process. Variables not included in this argument are ranked by their Variance Inflation Factor. See [preference_order()]. Default: NULL
-#' @param f (optional: function) Function to compute preference order. If NULL (default), the output of [f_default()] for the given data is used:
+#' @param preference_f (optional: function) Function to compute preference order. If NULL (default), the output of [f_default()] for the given data is used:
 #' \itemize{
 #'   \item [f_auc_rf()]: `response` is binomial.
 #'   \item [f_r2_pearson()]: `response` and `predictors` are numeric.
@@ -66,11 +68,17 @@
 #'   \item [f_r2_rf()]: in all other cases.
 #' }
 #' Default: NULL
+#' @param preference_warn_limit (optional, numeric) Preference value (R-squared, AUC, or Cramer's V) over which a warning flagging suspicious predictors is issued. Disabled if NULL. Default: 0.8
 #' @param cor_method (optional; character string) Method used to compute pairwise correlations. Either "pearson" or "spearman". Default: "pearson".
-#' @param max_cor (optional; numeric) Maximum correlation allowed between any pair of variables in `predictors`. Recommended values are between 0.5 and 0.9. Higher values return larger number of predictors with a higher multicollinearity. If NULL, the pairwise correlation analysis is disabled. Default: `0.75`
-#' @param max_vif (optional, numeric) Maximum Variance Inflation Factor allowed during variable selection. Recommended values are between 2.5 and 10. Higher values return larger number of predictors with a higher multicollinearity. If NULL, the variance inflation analysis is disabled. Default: 5.
+#' @param cor_max (optional; numeric) Maximum correlation allowed between any pair of variables in `predictors`. Recommended values are between 0.5 and 0.9. Higher values return larger number of predictors with a higher multicollinearity. If NULL, the pairwise correlation analysis is disabled. Default: `0.75`
+#' @param vif_max (optional, numeric) Maximum Variance Inflation Factor allowed during variable selection. Recommended values are between 2.5 and 10. Higher values return larger number of predictors with a higher multicollinearity. If NULL, the variance inflation analysis is disabled. Default: 5.
 #' @param quiet (optional; logical) If FALSE, messages generated during the execution of the function are printed to the console Default: FALSE
-#' @return character vector; names of selected predictors
+#'
+#' @return
+#' \itemize{
+#'   \item character vector if `response` is NULL or is a string.
+#'   \item named list if `response` is a character vector.
+#' }
 #'
 #' @examples
 #' #parallelization setup
@@ -101,13 +109,13 @@
 #' x <- collinear(
 #'   df = df,
 #'   predictors = predictors,
-#'   max_cor = 0.75, #default
-#'   max_vif = 5    #default
+#'   cor_max = 0.75, #default
+#'   vif_max = 5    #default
 #'   )
 #'
 #' x
 #'
-#' #all correlations below max_cor
+#' #all correlations below cor_max
 #' cor_df(
 #'   df = df,
 #'   predictors = x
@@ -129,7 +137,7 @@
 #' # x <- collinear(
 #' #   df = df,
 #' #   predictors = predictors,
-#' #   max_cor = NULL
+#' #   cor_max = NULL
 #' # )
 #'
 #' #correlation filtering only
@@ -140,7 +148,7 @@
 #' # x <- collinear(
 #' #   df = df,
 #' #   predictors = predictors,
-#' #   max_vif = NULL
+#' #   vif_max = NULL
 #' # )
 #'
 #'
@@ -247,12 +255,13 @@ collinear <- function(
     df = NULL,
     response = NULL,
     predictors = NULL,
-    encoding_method = "mean",
+    encoding_method = "loo",
     preference_order = NULL,
-    f = NULL,
+    preference_f = NULL,
+    preference_warn_limit = 0.8,
     cor_method = "pearson",
-    max_cor = 0.75,
-    max_vif = 5,
+    cor_max = 0.75,
+    vif_max = 5,
     quiet = FALSE
 ){
 
@@ -267,127 +276,199 @@ collinear <- function(
     quiet = quiet
   )
 
-  #validate response
-  response <- validate_response(
-    df = df,
-    response = response,
-    quiet = quiet
+  #managing multiple responses
+  responses <- intersect(
+    x = colnames(df),
+    y = response
   )
+  rm(response)
 
-  #validate predictors
-  predictors <- validate_predictors(
-    df = df,
-    response = response,
-    predictors = predictors
-  )
+  #copy of preference order
+  preference_order_user <- preference_order
 
-  # target encoding ----
-  #ignored if:
-  # response == NULL
-  # encoding_method == NULL
-  df <- target_encoding_lab(
-    df = df,
-    response = response,
-    predictors = predictors,
-    methods = encoding_method[1],
-    overwrite = TRUE,
-    quiet = quiet
-  )
+  #copy of predictors
+  predictors_user <- predictors
 
-  # preference order ----
-  if(is.null(preference_order)){
+  #output list
+  out <- list()
 
-    preference_order <- preference_order(
+  #iterate over responses
+  #runs even when responses is NULL
+  for(
+    response in
+    if (is.null(responses)) list(NULL) else responses
+    ){
+
+    if(is.list(response)){
+      response <- NULL
+    }
+
+    #validate response
+    response <- validate_response(
+      df = df,
+      response = response,
+      quiet = quiet
+    )
+
+    #if several responses, none can be null
+    if(length(responses) > 1){
+
+      if(is.null(response)){
+
+        if(quiet == FALSE){
+
+          message("collinear::collinear(): response '", response, "' is not valid, skipping it." )
+
+        }
+
+        next
+
+      } else {
+
+        if(quiet == FALSE){
+
+          message("collinear::collinear(): multicollinearity filtering for response: '", response, "'." )
+
+        }
+
+      }
+
+    }
+
+    #reset and validate predictors
+    predictors <- validate_predictors(
+      df = df,
+      response = response,
+      predictors = predictors_user
+    )
+
+    # target encoding ----
+    #ignored if:
+    # response == NULL
+    # encoding_method == NULL
+    df <- target_encoding_lab(
       df = df,
       response = response,
       predictors = predictors,
-      f = NULL,
+      methods = encoding_method[1],
+      overwrite = TRUE,
       quiet = quiet
-    )$predictor
+    )
 
-  } else if (
-    is.data.frame(preference_order) &&
-    "predictor" %in% names(preference_order)
-  ){
+    # preference order ----
+    if(is.null(preference_order_user)){
 
-    preference_order <- preference_order$predictor
+      preference_order <- preference_order(
+        df = df,
+        response = response,
+        predictors = predictors,
+        f = preference_f,
+        quiet = quiet,
+        warn_limit = preference_warn_limit
+      )$predictor
 
-  }
+    } else if (
+      is.data.frame(preference_order_user) &&
+      "predictor" %in% names(preference_order_user)
+    ){
 
-  # pairwise correlation filter ----
-  selection <- cor_select(
-    df = df,
-    predictors = predictors,
-    preference_order = preference_order,
-    cor_method = cor_method,
-    max_cor = max_cor,
-    quiet = quiet
-  )
+      preference_order <- preference_order_user$predictor
 
-  #vif filter
+    } else if (is.vector(preference_order_user)) {
 
-  #initialize selection.numeric
-  selection.numeric <- selection
-  selection.categorical <- NULL
+      preference_order <- preference_order_user
 
-  #if cor_select() filtered predictors
-  if(is.null(max_cor) == FALSE){
+    }
 
-    #separate numeric and categorical
-    selection.type <- identify_predictors(
+    # pairwise correlation filter ----
+    selection <- cor_select(
       df = df,
-      predictors = selection
+      predictors = predictors,
+      preference_order = preference_order,
+      cor_method = cor_method,
+      cor_max = cor_max,
+      quiet = quiet
     )
 
-    #store numeric and categorical
-    selection.numeric <- selection.type$numeric
-    selection.categorical <- selection.type$categorical
+    #vif filter
 
-  }
+    #initialize selection.numeric
+    selection.numeric <- selection
+    selection.categorical <- NULL
 
-  #run vif filtering
-  selection.vif <- vif_select(
-    df = df,
-    predictors = selection.numeric,
-    preference_order = preference_order,
-    max_vif = max_vif,
-    quiet = quiet
-  )
+    #if cor_select() filtered predictors
+    if(is.null(cor_max) == FALSE){
 
-  #merge selections
-  selection <- c(
-    selection.vif,
-    selection.categorical
-  ) |>
-    unique() |>
-    na.omit()
+      #separate numeric and categorical
+      selection.type <- identify_predictors(
+        df = df,
+        predictors = selection
+      )
 
-  #order as in preference order
-  if(all(selection %in% preference_order)){
-    selection <- selection[order(match(selection, preference_order))]
-  }
+      #store numeric and categorical
+      selection.numeric <- selection.type$numeric
+      selection.categorical <- selection.type$categorical
 
-  #message if vif_select() did nothing
-  if(
-    (
-      !is.null(selection.categorical) &&
-      !all(selection.vif %in% selection)
+    }
+
+    #run vif filtering
+    selection.vif <- vif_select(
+      df = df,
+      predictors = selection.numeric,
+      preference_order = preference_order,
+      vif_max = vif_max,
+      quiet = quiet
+    )
+
+    #merge selections
+    selection <- c(
+      selection.vif,
+      selection.categorical
+    ) |>
+      unique() |>
+      na.omit()
+
+    #order as in preference order
+    if(all(selection %in% preference_order)){
+
+      selection <- selection[order(match(selection, preference_order))]
+
+    }
+
+    #message with selection if required
+    if(
+      (
+        !is.null(selection.categorical) &&
+        !all(selection.vif %in% selection)
       ) &&
-    quiet == FALSE
-  ){
+      quiet == FALSE
+    ){
 
-    message(
-      "\ncollinear::collinear(): selected predictors: \n - ",
-      paste(selection, collapse = "\n - ")
-    )
+      message(
+        "\ncollinear::collinear(): selected predictors: \n - ",
+        paste(selection, collapse = "\n - ")
+      )
 
+    }
+
+    attr(
+      x = selection,
+      which = "validated"
+    ) <- TRUE
+
+    if(!is.null(response)){
+      out[[response]] <- selection
+    } else {
+      out[[1]] <- selection
+    }
+
+  } #end of loop
+
+  if(is.list(out) && length(out) == 1){
+    out <- unlist(out)
+    names(out) <- NULL
   }
 
-  attr(
-    x = selection,
-    which = "validated"
-  ) <- TRUE
-
-  selection
+  out
 
 }