+++ R/cor_clusters.R
@@ -2,7 +2,7 @@
 #'
 #' @description
 #'
-#' Hierarchical clustering of predictors from their pairwise correlation matrix. Computes the correlation matrix with [cor_df()] and [cor_matrix()], transforms it to a dist object, computes a clustering solution with [stats::hclust()], and applies [stats::cutree()] to separate groups based on the value of the argument `max_cor`.
+#' Hierarchical clustering of predictors from their pairwise correlation matrix. Computes the correlation matrix with [cor_df()] and [cor_matrix()], transforms it to a dist object, computes a clustering solution with [stats::hclust()], and applies [stats::cutree()] to separate groups based on the value of the argument `cor_max`.
 #'
 #' Returns a data frame with predictor names and their clusters, and optionally, prints a dendrogram of the clustering solution.
 #'
@@ -39,7 +39,7 @@ cor_clusters <- function(
     df = NULL,
     predictors = NULL,
     cor_method = "pearson",
-    max_cor = 0.75,
+    cor_max = 0.75,
     method = "complete",
     plot = FALSE
 ){
@@ -59,7 +59,7 @@ cor_clusters <- function(
 
   hc_groups <- stats::cutree(
     tree = hc,
-    h = 1 - max_cor,
+    h = 1 - cor_max,
   )
 
   df_clusters <- data.frame(
@@ -74,7 +74,7 @@ cor_clusters <- function(
     plot(
       x = hc,
       labels = df_clusters$predictor,
-      main = paste0("Clustered predictors - max_cor = ", max_cor),
+      main = paste0("Clustered predictors - cor_max = ", cor_max),
       xlab = "",
       ylab = "1 - Cor",
       sub = "",
@@ -83,7 +83,7 @@ cor_clusters <- function(
     )
 
     graphics::abline(
-      h = 1 - max_cor,
+      h = 1 - cor_max,
       col = "gray50",
       lty = 3,
       lwd = 2
@@ -91,7 +91,7 @@ cor_clusters <- function(
 
     stats::rect.hclust(
       tree = hc,
-      h = 1 - max_cor,
+      h = 1 - cor_max,
       border = "red4"
     )