+++ dev_scripts/microbenchmark.R
@@ -39,7 +39,7 @@ option1 <- function() {
 
   for(i in preference_order_rev){
 
-    if(max(m[selected, i]) > max_cor){
+    if(max(m[selected, i]) > cor_max){
       selected <- selected[selected != i]
     }
 
@@ -88,8 +88,8 @@ option2 <- function() {
 
     i <- i + 1
 
-    #remove i column and row if max > max_cor
-    if(max(m[1:i, i]) > max_cor){
+    #remove i column and row if max > cor_max
+    if(max(m[1:i, i]) > cor_max){
 
       #remove column
       m <- m[-i, -i, drop = FALSE]